chore: migrate LayoutItemAnimator
