+++ src/components/asset-list/RecyclerAssetList/Shared.ts
@@ -1,0 +1,15 @@
+export type Shared = {
+  globalDeviceDimensions: number;
+};
+
+export default ((): Shared => {
+  let globalDeviceDimensions = 0;
+  return {
+    get globalDeviceDimensions(): number {
+      return globalDeviceDimensions;
+    },
+    set globalDeviceDimensions(nextGlobalDeviceDimensions: number) {
+      globalDeviceDimensions = nextGlobalDeviceDimensions;
+    },
+  };
+})();
